backend/requirements.txt
# Core
fastapi==0.115.*
uvicorn[standard]==0.34.*
python-dotenv==1.1.*
pydantic==2.11.*

# HTTP client
httpx==0.28.*

# Knowledge graph
networkx==3.4.*

# LLM APIs
anthropic==0.52.*
google-genai==1.14.*

# Database
aiosqlite==0.21.*

# Geospatial
geopy==2.4.*

# MCP server
fastmcp>=2.0.0

# Testing
pytest==8.3.*
pytest-asyncio==0.25.*
